(#97) moving LinkBuilder to the domain's logic
